# tests/Sample/OpenAPIs/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "Naos",
    "description": "Naos",
    "termsOfService": "None",
    "contact": {
      "name": "Naos",
      "url": "https://github.com/vip32/Naos.Core",
      "email": ""
    },
    "version": "v1"
  },
  "host": "localhost:5001",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/operations/logevents": {
      "get": {
        "tags": [
          "Naos Operations"
        ],
        "operationId": "NaosOperationsLogEvents_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LogEvent"
              }
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/operations/logevents/dashboard": {
      "get": {
        "tags": [
          "Naos Operations"
        ],
        "operationId": "NaosOperationsLogEvents_GetHtml",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProblemDetails": {
      "type": "object",
      "description": "A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807.",
      "additionalProperties": {},
      "properties": {
        "type": {
          "type": "string",
          "description": "A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when\ndereferenced, it provide human-readable documentation for the problem type\n(e.g., using HTML [W3C.REC-html5-20141028]).  When this member is not present, its value is assumed to be\n\"about:blank\"."
        },
        "title": {
          "type": "string",
          "description": "A short, human-readable summary of the problem type.It SHOULD NOT change from occurrence to occurrence\nof the problem, except for purposes of localization(e.g., using proactive content negotiation;\nsee[RFC7231], Section 3.4)."
        },
        "status": {
          "type": "integer",
          "description": "The HTTP status code([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.",
          "format": "int32"
        },
        "detail": {
          "type": "string",
          "description": "A human-readable explanation specific to this occurrence of the problem."
        },
        "instance": {
          "type": "string",
          "description": "A URI reference that identifies the specific occurrence of the problem.It may or may not yield further information if dereferenced."
        }
      }
    },
    "ValidationProblemDetails": {
      "allOf": [
        {
          "$ref": "#/definitions/ProblemDetails"
        },
        {
          "type": "object",
          "description": "A ProblemDetails for validation errors.",
          "additionalProperties": {},
          "properties": {
            "errors": {
              "type": "object",
              "description": "Gets the validation errors associated with this instance of ValidationProblemDetails.",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      ]
    },
...
    "LogEvent": {
      "type": "object",
      "required": [
        "timestamp",
        "ticks"
      ],
      "properties": {
        "level": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "environment": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "ticks": {
          "type": "integer",
          "format": "int64"
        },
        "correlationId": {
          "type": "string"
        },
        "serviceName": {
          "type": "string"
        },
        "serviceProduct": {
          "type": "string"
        },
        "serviceCapability": {
          "type": "string"
        },
        "sourceContext": {
          "type": "string"
        },
        "trackType": {
          "type": "string"
        },
        "trackId": {
          "type": "string"
        },
        "properties": {
          "type": "object",
          "additionalProperties": {}
        }
      }
    },
...
  }
}